requirements.txt
flask==0.11.1
Flask-SQLAlchemy==2.1
requests==2.12.4
